Dedupe highlight-adjacent code: shared mdPunct, useDismissable, derived color rows
Post-review cleanups on the highlights plugin (no behavior change):

- Promote the `.md-punct` reveal-fence El builder to `mdPunct()` in
  inline-code.ts; emphasis and highlight both call it instead of each
  re-declaring the identical closure.
- Extract the popover dismiss effect (deferred outside-pointerdown + Escape)
  to `useDismissable(ref, onClose)`; adopt it in the highlight color menu,
  the tag color menu, and the link-edit popover — three byte-identical
  copies collapse to one.
- Derive the highlight color menu's `ROWS` from `HIGHLIGHT_EMOJI` (the one
  palette source of truth) and drop the redundant `ariaLabel` field, so a
  new palette color can't silently miss the menu.

typecheck + lint + unit tests green. e2e not run locally (chrome-headless-shell
fails to launch in this environment); changes are output-/logic-identical.
